+++ Controllers/TareaController.cs
@@ -2,6 +2,7 @@ using System.Diagnostics;
 using Microsoft.AspNetCore.Mvc;
 using  tl2_tp10_2023_IvanDMir.Models;
 using tl2_tp10_2023_IvanDMir.repositorios;
+using tl2_tp10_2023_IvanDMir.ViewModels;
 
 namespace tl2_tp10_2023_IvanDMir.Controllers;
 
@@ -16,36 +17,63 @@ public class TareaController : Controller
         repo = new TareaRepositorios();
     }
 
-        public IActionResult Index()
-    {
-        List<Tarea> Tareas = repo.GetAll();
-        return View(Tareas);
-    
+  [HttpGet]
+    public IActionResult Index() {
+        if(!isLogin()) return RedirectToRoute(new { controller = "Login", action = "Index"});
+        if(esAdmin()) return View(new GTViewModel(repo.GetAll()));
+        var loggedUserId = Convert.ToInt32(HttpContext.Session.GetString("Id"));
+        return View(new GTViewModel(repo.GetByUser(loggedUserId)));
     }
+ [HttpGet]
+    public IActionResult Add() => View(new ATViewModel());
 
-     [HttpGet]
-    public IActionResult Add(){
-        return View(new Tarea());
-    }
     [HttpPost]
-    public IActionResult Add(Tarea tab){
-        repo.Crear(tab);
-         return RedirectToAction("Index");
-    }
-     [HttpGet]
-    public IActionResult Update(int id){
-        return View(repo.GetById(id));
-    }
-    [HttpPost]
-    public IActionResult Update(Tarea tarea){
-        repo.Update(tarea.Id,tarea);
+    public IActionResult Add(ATViewModel tarea) {
+        if(!ModelState.IsValid) return RedirectToAction("Index");
+        var nuevaTarea = new Tarea() {
+            Nombre = tarea.Nombre,
+            Descripcion = tarea.Descripcion,
+            Estado = Estados.ToDo,
+            Color = tarea.Color,
+            IdTablero = tarea.IdTablero
+        };
+        repo.Crear(nuevaTarea);
         return RedirectToAction("Index");
-        
-
     }
+   [HttpGet]
+    public IActionResult Update(int id) => View(new UTViewModel(repo.GetById(id)));
+
+    [HttpPost]
+    public IActionResult Update(UTViewModel tarea) {
+        if(!ModelState.IsValid) return RedirectToAction("Index");
+        var ttarea = repo.GetAll().FirstOrDefault(t => t.Id == tarea.Id);
+        repo.Update(ttarea.Id, ttarea);
+        return RedirectToAction("Index");
+    }
+
+
     [HttpGet]
-    public IActionResult Delete(int id){  
+    public IActionResult Delete(int id) {
+        if(!ModelState.IsValid) return RedirectToAction("Index");
         repo.Delete(id);
-      return RedirectToAction("Index");
+        return RedirectToAction("Index");
+    }
+       private bool isLogin()
+    {
+        if (HttpContext.Session != null ){
+            return true;
+        }else{
+            return false;
+        }
+    }
+
+    private bool esAdmin(){
+         if (HttpContext.Session.GetString("Rol") == Enum.GetName(Roles.admin)){
+            return true;
+         }
+         var colo = HttpContext.Session.GetString("Rol");
+         var malo =  Enum.GetName(Roles.admin);
+         return false;
+        
     }
 }
